Refactoreret mapperne
