--- src/main/java/app/gruppe1Config/Gruppe1ThymeleafConfig.java
@@ -1,17 +1,0 @@
-package app.gruppe1Config;
-
-import org.thymeleaf.TemplateEngine;
-import org.thymeleaf.templateresolver.ClassLoaderTemplateResolver;
-
-public class Gruppe1ThymeleafConfig
-{
-    public static TemplateEngine templateEngine()
-    {
-        TemplateEngine templateEngine = new TemplateEngine();
-        ClassLoaderTemplateResolver templateResolver = new ClassLoaderTemplateResolver();
-        templateResolver.setPrefix("templates/"); // assuming templates are in resources/templates/
-        templateResolver.setSuffix(".html");
-        templateEngine.setTemplateResolver(templateResolver);
-        return templateEngine;
-    }
-}
