+++ src/main/java/app/persistence/gruppe1/Gruppe1ConnectionPool.java
@@ -1,4 +1,4 @@
-package app.gruppe1Persistence;
+package app.persistence.gruppe1;
 
 public class Gruppe1ConnectionPool
 {
